# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Cadastro de Usuários",
    "description": "Documentação das rotas de cadastro e consulta de instituições beneficiárias e pessoas físicas.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/login": {
      "post": {
        "summary": "Login de Usuário",
        "description": "Permite o login de uma pessoa beneficiária, instituição beneficiária ou doador, retornando um token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "userType"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "usuario@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "senha123"
                  },
                  "userType": {
                    "type": "string",
                    "enum": ["pessoa", "instituicao", "doador"],
                    "example": "pessoa"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tipo": {
                      "type": "string",
                      "example": "pessoa"
                    },
                    "id": {
                      "type": "integer",
                      "example": 12
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não verificado ou credenciais inválidas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Usuário ou senha inválidos para pessoa"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tipo de usuário não reconhecido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Tipo de usuário não reconhecido"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro no servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
"/recuperaremail/{tipo}/{email}": {
  "get": {
    "summary": "Recuperar dados por e-mail",
    "description": "Busca usuário pelo e-mail em uma das tabelas: instituição, doador ou pessoa beneficiária, conforme o tipo especificado.",
    "parameters": [
      {
        "name": "tipo",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": ["instituicao", "doador", "pessoa"]
        },
        "description": "Tipo do usuário para busca"
      },
      {
        "name": "email",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "email"
        },
        "description": "E-mail do usuário a ser buscado"
      }
    ],
    "responses": {
      "200": {
        "description": "Usuário encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Usuário encontrado"
                },
                "dados": {
                  "type": "object",
                  "description": "Dados do usuário encontrado",
                  "example": {
                    "id": 1,
                    "name": "Nome do Usuário",
                    "email": "email@exemplo.com"
                  }
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Tipo de usuário inválido"
      },
      "404": {
        "description": "E-mail não encontrado"
      },
      "500": {
        "description": "Erro no servidor"
      }
    }
  }
},
"/enviar-recuperacao": {
  "post": {
    "summary": "Enviar e-mail para recuperação de senha",
    "description": "Envia um e-mail com link para recuperação de senha ao usuário conforme o tipo informado.",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email",
                "example": "usuario@example.com"
              },
              "tipo": {
                "type": "string",
                "enum": ["instituicao", "doador", "pessoa"],
                "example": "doador"
              }
            },
            "required": ["email", "tipo"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "E-mail de recuperação enviado com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "E-mail de recuperação enviado com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Email e tipo são obrigatórios"
      },
      "500": {
        "description": "Erro ao enviar o e-mail de recuperação"
      }
    }
  }
  },
  "/atualizar-senha": {
  "put": {
    "summary": "Atualizar senha do usuário",
    "description": "Atualiza a senha do usuário conforme o tipo (instituição, doador ou pessoa). A senha é armazenada criptografada.",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email",
                "example": "usuario@example.com"
              },
              "tipo": {
                "type": "string",
                "enum": ["instituicao", "doador", "pessoa"],
                "example": "pessoa"
              },
              "novaSenha": {
                "type": "string",
                "example": "NovaSenha123"
              }
            },
            "required": ["email", "tipo", "novaSenha"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Senha atualizada com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Senha atualizada com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Email, tipo ou nova senha ausentes ou tipo inválido"
      },
      "404": {
        "description": "Usuário não encontrado"
      },
      "500": {
        "description": "Erro no servidor ao atualizar a senha"
      }
    }
  }
  },
...
}
}